+++ src/components/AddPlayer.vue
@@ -1,23 +1,41 @@
 <script>
 
+import {db, push, ref} from "@/assets/js/firebase";
+
 export default {
-  data() {
-    return {
-      messageSpotify: 'Se connecter à Spotify'
-    }
-  },
-  props : {
-    users : Object,
-    showMenuAddPlayer : Boolean
-  },
+  methods: {
+    // Ajout d'un utilisateur dans la bdd
+    addUser() {
+      // Récupération des infos de l'utilisateur
+      let username = document.querySelector('#username').value;
+      // let buzzerId = document.querySelector('#buzzer').value;
+
+      // Génération d'un id pour le state pour l'api Dicebear
+      let id = new Date();
+      id = id.getTime();
+
+      // Récupération d'un avatar random
+      let profilePicture = 'https://avatars.dicebear.com/api/avataaars/' + id + '.svg'
+
+      // On push vers firebase le nouvel utilisateur
+      push(ref(db, import.meta.env.VITE_FIREBASE_DB_USERS), {
+        "buzzerId": 0,
+        "username": username,
+        "profilePicture": profilePicture,
+        "points": 0
+      }).key;
+
+      // On cache le menu d'ajout d'un nouvel utilisateur
+      this.$store.commit('setShowMenuAddPlayer', false)
+    },
+  }
 }
 </script>
 <template>
-  <div class="card" :style="showMenuAddPlayer ? 'visibility: visible;' : 'visibility: hidden;'">
+  <div class="card" :style="this.$store.getters.getShowMenuAddPlayer ? 'visibility: visible;' : 'visibility: hidden;'" style="width: 100%; margin-top: 3%">
     <div class="addUser flex space-around">
       <input id="username" placeholder="Pseudo">
-<!--      <input id="buzzer" placeholder="N° du Buzzer">-->
-      <span id="addUser" class="button" @click="$emit('addUser')">Ajouter un utilisateur</span>
+      <span id="addUser" class="button" @click="addUser">Ajouter un utilisateur</span>
     </div>
   </div>
 </template>