+++ src/components/Modal.vue
@@ -1,21 +1,18 @@
 <script>
-import {db, ref, set, onValue} from '../assets/js/firebase.js';
+import {db, ref, set, onValue} from '@/assets/js/firebase';
 
 export default {
   data() {
     return {
-      open: false
     }
   },
   props : {
-    clicker : String,
-    isMusicPlaying : Boolean,
     points : Number
   },
   watch : {
-    clicker() {
-      if (this.isMusicPlaying){
-        this.open = true;
+    'getClicker'(clicker, previousClicker) {
+      if (this.$store.getters.getIsMusicPlaying && previousClicker !== clicker && clicker !== '' && (!this.$store.getters.getArtist || !this.$store.getters.getTitle)){
+        this.$store.dispatch('changeModalStatus', {isShowModal : true, blur: 60, isMusicPlaying : false})
       }
     }
   },
@@ -29,7 +26,7 @@ export default {
       onValue(ref(db, userNode), (snapshot) => {
         snapshot.forEach((childSnapshot) => {
           let user = childSnapshot.val();
-          if (user.username === this.clicker){
+          if (user.username === this.getClicker){
             newPoints = parseInt(user.points) + parseInt(points);
             userKey = childSnapshot.key;
           }
@@ -38,20 +35,28 @@ export default {
 
       // On set le nouveau nombre de points
       set(ref(db, userNode+'/'+userKey + '/points'), newPoints);
+
+      this.$store.dispatch('changeModalStatus', {isShowModal: false, blur: 0, isMusicPlaying: true})
+    },
+  },
+  computed: {
+    getClicker(){
+      return this.$store.getters.getClicker;
     }
   }
 }
 </script>
 
 <template>
-  <div v-if="open" class="modal borderColored">
-    <span style="font-size: 2vw; cursor: pointer" @click="open = false; $emit('resetBlurAndWinner', false)"><font-awesome-icon icon="fa-solid fa-circle-xmark" /></span>
-    <p style="font-size: 2vw; text-align: center">{{clicker}} a buzzé !!</p>
+  <div v-if="this.$store.state.showModal" class="modal borderColored">
+    <span style="font-size: 2vw; cursor: pointer" @click="this.$store.dispatch('changeModalStatus', {isShowModal : false, blur: 0, isMusicPlaying : true})"><font-awesome-icon icon="fa-solid fa-circle-xmark" /></span>
+    <p style="font-size: 2vw; text-align: center">{{this.$store.getters.getClicker}} a buzzé !!</p>
 
     <div style="display: flex; justify-content: space-around; margin-top: 3vh">
-      <div class="button" @click="addPointsToUser(1); open = false; $emit('resetBlurAndWinner', false)" ><font-awesome-icon icon="fa-solid fa-microphone-lines"/> Artiste</div>
-      <div class="button" @click="addPointsToUser(1); open = false; $emit('resetBlurAndWinner', false)" ><font-awesome-icon icon="fa-solid fa-compact-disc"/> Titre</div>
-      <div class="button" @click="addPointsToUser(3); open = false; $emit('resetBlurAndWinner', true)" ><font-awesome-icon icon="fa-solid fa-music"/> Artiste et Titre</div>
+      <div :class="this.$store.getters.getArtist ? 'button-disabled button' : 'button'" @click="addPointsToUser(1); this.$store.commit('setArtist', true)" ><font-awesome-icon icon="fa-solid fa-microphone-lines"/> {{ this.$store.getters.getArtist ? 'Artiste trouvé' : 'Artiste'}}</div>
+      <div :class="this.$store.getters.getTitle ? 'button-disabled button' : 'button'" @click="addPointsToUser(1); this.$store.commit('setTitle', true)" ><font-awesome-icon icon="fa-solid fa-compact-disc"/>
+        {{ this.$store.getters.getTitle ? 'Titre trouvé' : 'Titre'}}</div>
+      <div class="button" @click="addPointsToUser(3); this.$store.commit('setTitle', true); this.$store.commit('setArtist', true)" ><font-awesome-icon icon="fa-solid fa-music"/> Artiste et Titre</div>
     </div>
   </div>
 </template>
@@ -65,6 +70,5 @@ export default {
   width: 50%;
   height: 33vh;
   padding: 2vh;
-  border-radius: 5%;
 }
 </style>