+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # CHANGELOG.md
 
+## 1.0.3 (2022-09-30)
+
+- Désactivation des boutons artiste/titre si déjà trouvé(s)
+- Refonte graphique
+- Utilisation de QrCode avec la [libraire suivante](https://github.com/scopewu/qrcode.vue)
+
+## 1.0.2 (2022-08-23)
+
+- Ajout d'un décompte, passage à la prochaine chanson quand décompte atteint
+- Changement de musique quand artiste et titre trouvé
+- Ajout d'un store Vuex
+
 ## 1.0.1 (2022-07-28)
 
 - Gestion des points